Check id on User endpoints

+++ api/src/main/java/com/daar/mylibrary/controller/UserController.java
@@ -4,10 +4,12 @@ import com.daar.mylibrary.dto.response.*;
 import com.daar.mylibrary.dto.response.Books.BooksShortResponse;
 import com.daar.mylibrary.exception.NotFoundException;
 import com.daar.mylibrary.service.UsersService;
+import com.daar.mylibrary.utils.Constants;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -18,6 +20,7 @@ import org.springframework.util.StopWatch;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Controller
@@ -33,8 +36,10 @@ public class UserController {
     @ApiResponse(responseCode = "401", description = "The authentication or authorization failed", content = { @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class)) })
     @ApiResponse(responseCode = "404", description = "User not founded", content = { @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class)) })
     @ApiResponse(responseCode = "500", description = "Internal error", content = { @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class)) })
+    @SecurityRequirement(name = "globalSecurity")
     @GetMapping("/{id}")
-    public ResponseEntity<Response> getUserSuggestion(@PathVariable String id) {
+    public ResponseEntity<Response> getUserSuggestion(@PathVariable String id, @RequestHeader(name = "Authorization", required = false) String token) {
+        if (!Objects.equals(id, Constants.decodeToken(token).sub)) return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ErrorResponse("User id does not match"));
         try {
             StopWatch watch = new StopWatch();
             watch.start();
@@ -60,8 +65,10 @@ public class UserController {
     @ApiResponse(responseCode = "200", description = "User added", content = { @Content(mediaType = "application/json", schema = @Schema(implementation = ElementRemovedResponse.class)) })
     @ApiResponse(responseCode = "404", description = "User not founded", content = { @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class)) })
     @ApiResponse(responseCode = "500", description = "Internal error", content = { @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class)) })
+    @SecurityRequirement(name = "globalSecurity")
     @PatchMapping("/{id}/read/{book_id}")
-    public ResponseEntity<Response> addUserBook(@PathVariable String id, @PathVariable(name = "book_id") String bookId) {
+    public ResponseEntity<Response> addUserBook(@PathVariable String id, @PathVariable(name = "book_id") String bookId, @RequestHeader(name = "Authorization", required = false) String token) {
+        if (!Objects.equals(id, Constants.decodeToken(token).sub)) return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ErrorResponse("User id does not match"));
         try {
             return ResponseEntity.status(HttpStatus.OK).body(new UserResponse(usersService.addBooksRead(id, bookId)));
         } catch (NotFoundException e) {
@@ -75,8 +82,10 @@ public class UserController {
     @ApiResponse(responseCode = "200", description = "User added", content = { @Content(mediaType = "application/json", schema = @Schema(implementation = ElementRemovedResponse.class)) })
     @ApiResponse(responseCode = "404", description = "User not founded", content = { @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class)) })
     @ApiResponse(responseCode = "500", description = "Internal error", content = { @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class)) })
+    @SecurityRequirement(name = "globalSecurity")
     @PatchMapping("/{id}/keywords")
-    public ResponseEntity<Response> addUserSearch(@PathVariable String id, @RequestBody List<String> keywords) {
+    public ResponseEntity<Response> addUserSearch(@PathVariable String id, @RequestBody List<String> keywords, @RequestHeader(name = "Authorization", required = false) String token) {
+        if (!Objects.equals(id, Constants.decodeToken(token).sub)) return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ErrorResponse("User id does not match"));
         try {
             return ResponseEntity.status(HttpStatus.OK).body(new UserResponse(usersService.addKeyword(id, keywords)));
         } catch (NotFoundException e) {
@@ -90,8 +99,10 @@ public class UserController {
     @ApiResponse(responseCode = "200", description = "User removed", content = { @Content(mediaType = "application/json", schema = @Schema(implementation = ElementRemovedResponse.class)) })
     @ApiResponse(responseCode = "404", description = "User not founded", content = { @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class)) })
     @ApiResponse(responseCode = "500", description = "Internal error", content = { @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class)) })
+    @SecurityRequirement(name = "globalSecurity")
     @DeleteMapping("/{id}")
-    public ResponseEntity<Response> deleteUser(@PathVariable String id) {
+    public ResponseEntity<Response> deleteUser(@PathVariable String id, @RequestHeader(name = "Authorization", required = false) String token) {
+        if (!Objects.equals(id, Constants.decodeToken(token).sub)) return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ErrorResponse("User id does not match"));
         try {
             return ResponseEntity.status(HttpStatus.OK).body(new UserResponse(usersService.deleteUser(id)));
         } catch (NotFoundException e) {
